+++ Alignment/src/alignment/Local.java
@@ -11,9 +11,9 @@ public class Local extends Alignment{
 	}
 
 	public void calcMatrix(int i, int j) {		
-				float gap1 = aMatrix[i][j-1] + gapFunction.calcPenalty(1);
-				float gap2 = aMatrix[i-1][j] + gapFunction.calcPenalty(1);
-				float match = aMatrix[i-1][j-1] + scoringMatrix.getScore(sequence.getSequenceA()[i-1], sequence.getSequenceB()[j-1]);
+				int gap1 = aMatrix[i][j-1] + gapFunction.calcPenalty(1);
+				int gap2 = aMatrix[i-1][j] + gapFunction.calcPenalty(1);
+				int match = aMatrix[i-1][j-1] + scoringMatrix.getScore(sequence.getSequenceA()[i-1], sequence.getSequenceB()[j-1]);
 				aMatrix[i][j] = Math.max(0, Math.max(match, Math.max(gap1, gap2)));
 		
 	}
